improved session action logs
